[3787] Move classes from ui-common to ui-components
