--- stream-chat-android-ui-common/src/main/kotlin/com/getstream/sdk/chat/utils/strings/ChatStrings.kt
@@ -1,24 +1,0 @@
-/*
- * Copyright (c) 2014-2022 Stream.io Inc. All rights reserved.
- *
- * Licensed under the Stream License;
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    https://github.com/GetStream/stream-chat-android/blob/main/LICENSE
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.getstream.sdk.chat.utils.strings
-
-import androidx.annotation.StringRes
-
-public interface ChatStrings {
-    public fun get(@StringRes resId: Int): String
-    public fun get(@StringRes resId: Int, vararg formatArgs: Any?): String
-}
