+++ stream-chat-android-ui-common/src/test/kotlin/com/getstream/sdk/chat/Mother.kt
@@ -39,44 +39,6 @@ import java.io.File
 import java.time.Instant
 import java.util.Date
 
-internal fun randomUser(
-    id: String = randomString(),
-    name: String = randomString(),
-    image: String = randomString(),
-    role: String = randomString(),
-    invisible: Boolean = randomBoolean(),
-    banned: Boolean = randomBoolean(),
-    devices: List<Device> = mutableListOf(),
-    online: Boolean = randomBoolean(),
-    createdAt: Date? = null,
-    updatedAt: Date? = null,
-    lastActive: Date? = null,
-    totalUnreadCount: Int = positiveRandomInt(),
-    unreadChannels: Int = positiveRandomInt(),
-    mutes: List<Mute> = mutableListOf(),
-    teams: List<String> = listOf(),
-    channelMutes: List<ChannelMute> = emptyList(),
-    extraData: MutableMap<String, Any> = mutableMapOf()
-): User = User(
-    id,
-    name,
-    image,
-    role,
-    invisible,
-    banned,
-    devices,
-    online,
-    createdAt,
-    updatedAt,
-    lastActive,
-    totalUnreadCount,
-    unreadChannels,
-    mutes,
-    teams,
-    channelMutes,
-    extraData
-)
-
 internal fun createUser(
     id: String = randomString(),
     name: String = randomString(),
