+++ stream-chat-android-ui-components/src/main/kotlin/com/getstream/sdk/chat/utils/GridSpacingItemDecoration.java
